feat(cache): add cache to the repository

+++ lib/domain/data/model/monitoring_data.dart
@@ -1,13 +1,13 @@
 class MonitoringData {
   final DateTime timestamp;
-  final double value;
+  final int value;
 
   MonitoringData({required this.timestamp, required this.value});
 
   factory MonitoringData.fromJson(Map<String, dynamic> json) {
     return MonitoringData(
       timestamp: DateTime.parse(json['timestamp']),
-      value: json['value'].toDouble(),
+      value: json['value'],
     );
   }
 }
